requirements.txt
PyQt5
yt-dlp
python-vlc